Created images in resource folder to simulate map preview, created working slider to scroll through various maps, fixed comments, moved gui to view partition
